requirements.txt
grpcio
grpcio-reflection
protobuf
numpy
opencv-python-headless
matplotlib
torch
torchvision
Pillow
mediapipe
scipy